+++ RWAPIMicroservicePython/errors.py
@@ -17,3 +17,12 @@ class NotFound(Error):
     pass
 
 
+class ApiKeyError(Error):
+    pass
+
+
+class ValidationError(Error):
+
+    def __init__(self, message, code):
+        super().__init__(message)
+        self.code = code