+++ RWAPIMicroservicePython/tests/cors_tests.py
@@ -1,6 +1,6 @@
 import json
-
 import requests_mock
+from RWAPIMicroservicePython.tests.mocks import mock_request_validation
 from flask import Flask, Blueprint, request
 
 import RWAPIMicroservicePython
@@ -8,6 +8,7 @@ import RWAPIMicroservicePython
 
 @requests_mock.mock(kw='mocker')
 def test_cors_headers_are_present(mocker):
+    get_user_data_calls = mock_request_validation(mocker, user=None)
     test_endpoints = Blueprint('rw_api', __name__)
 
     @test_endpoints.route('/test', methods=['GET', 'PUT', 'POST', 'PATCH', 'DELETE'])
@@ -23,10 +24,11 @@ def test_cors_headers_are_present(mocker):
     RWAPIMicroservicePython.register(
         app=app,
         gateway_url='http://gateway-url.com',
-        token='microserviceToken'
+        token='microserviceToken',
+        require_api_key=True
     )
 
-    response = app.test_client().get('/test')
+    response = app.test_client().get('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
     assert response.headers['access-control-allow-origin'] == '*'
@@ -34,7 +36,7 @@ def test_cors_headers_are_present(mocker):
     assert response.headers['access-control-allow-credentials'] == 'true'
     assert response.headers['access-control-allow-methods'] == 'OPTIONS,GET,PUT,POST,PATCH,DELETE'
 
-    response = app.test_client().put('/test')
+    response = app.test_client().put('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
     assert response.headers['access-control-allow-origin'] == '*'
@@ -42,7 +44,7 @@ def test_cors_headers_are_present(mocker):
     assert response.headers['access-control-allow-credentials'] == 'true'
     assert response.headers['access-control-allow-methods'] == 'OPTIONS,GET,PUT,POST,PATCH,DELETE'
 
-    response = app.test_client().post('/test')
+    response = app.test_client().post('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
     assert response.headers['access-control-allow-origin'] == '*'
@@ -50,7 +52,7 @@ def test_cors_headers_are_present(mocker):
     assert response.headers['access-control-allow-credentials'] == 'true'
     assert response.headers['access-control-allow-methods'] == 'OPTIONS,GET,PUT,POST,PATCH,DELETE'
 
-    response = app.test_client().patch('/test')
+    response = app.test_client().patch('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
     assert response.headers['access-control-allow-origin'] == '*'
@@ -58,7 +60,7 @@ def test_cors_headers_are_present(mocker):
     assert response.headers['access-control-allow-credentials'] == 'true'
     assert response.headers['access-control-allow-methods'] == 'OPTIONS,GET,PUT,POST,PATCH,DELETE'
 
-    response = app.test_client().delete('/test')
+    response = app.test_client().delete('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
     assert response.headers['access-control-allow-origin'] == '*'