+++ RWAPIMicroservicePython/tests/inject_logged_user_tests.py
@@ -1,5 +1,6 @@
 import json
-
+from RWAPIMicroservicePython.tests.mocks import mock_request_validation, mock_request_validation_invalid_token
+from RWAPIMicroservicePython.tests.constants import USER
 import requests_mock
 from flask import Flask, Blueprint, request
 
@@ -8,46 +9,18 @@ import RWAPIMicroservicePython
 
 @requests_mock.mock(kw='mocker')
 def test_inject_logged_user(mocker):
-    get_user_data_calls = mocker.get('http://gateway-url.com/auth/user/me',
-                                     request_headers={'Authorization': 'Bearer abcd'},
-                                     status_code=200,
-                                     json={
-                                         'id': '1a10d7c6e0a37126611fd7a7',
-                                         'name': 'test admin',
-                                         'role': 'ADMIN',
-                                         'provider': 'local',
-                                         'email': 'user@control-tower.org',
-                                         'extraUserData': {
-                                             'apps': [
-                                                 'rw',
-                                                 'gfw',
-                                                 'gfw-climate',
-                                                 'prep',
-                                                 'aqueduct',
-                                                 'forest-atlas',
-                                                 'data4sdgs'
-                                             ]
-                                         }
-                                     })
+    get_user_data_calls = mock_request_validation(mocker)
 
     test_endpoints = Blueprint('rw_api', __name__)
 
     @test_endpoints.route('/test', methods=['GET', 'PUT', 'POST', 'PATCH', 'DELETE'])
     def test_route():
         logged_user = json.loads(request.args.get("loggedUser", '{}'))
-        assert logged_user.get('name') == 'test admin'
-        assert logged_user.get('role') == 'ADMIN'
-        assert logged_user.get('provider') == 'local'
-        assert logged_user.get('email') == 'user@control-tower.org'
-        assert logged_user.get('extraUserData').get('apps') == [
-            'rw',
-            'gfw',
-            'gfw-climate',
-            'prep',
-            'aqueduct',
-            'forest-atlas',
-            'data4sdgs'
-        ]
+        assert logged_user.get('name') == USER['name']
+        assert logged_user.get('role') == USER['role']
+        assert logged_user.get('provider') == USER['provider']
+        assert logged_user.get('email') == USER['email']
+        assert logged_user.get('extraUserData').get('apps') == USER['extraUserData']['apps']
         return 'ok', 200
 
     app = Flask(__name__)
@@ -57,26 +30,27 @@ def test_inject_logged_user(mocker):
     RWAPIMicroservicePython.register(
         app=app,
         gateway_url='http://gateway-url.com',
-        token='microserviceToken'
+        token='microserviceToken',
+        require_api_key=True
     )
 
-    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
@@ -87,7 +61,7 @@ def test_inject_logged_user(mocker):
 @requests_mock.mock(kw='mocker')
 def test_inject_logged_user_when_no_authorization_header_is_present(mocker):
     # This is never actually called, were just using it as a way to validate that no calls are made to this endpoint
-    get_user_data_calls = mocker.get('http://gateway-url.com/auth/user/me', status_code=200, json={})
+    get_user_data_calls = mock_request_validation(mocker, user=None, application=None)
 
     test_endpoints = Blueprint('rw_api', __name__)
 
@@ -104,46 +78,37 @@ def test_inject_logged_user_when_no_authorization_header_is_present(mocker):
     RWAPIMicroservicePython.register(
         app=app,
         gateway_url='http://gateway-url.com',
-        token='microserviceToken'
+        token='microserviceToken',
+        require_api_key=True
     )
 
-    response = app.test_client().get('/test')
+    response = app.test_client().get('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().put('/test')
+    response = app.test_client().put('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().post('/test')
+    response = app.test_client().post('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().patch('/test')
+    response = app.test_client().patch('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().delete('/test')
+    response = app.test_client().delete('/test', headers={'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    assert not get_user_data_calls.called
-    assert get_user_data_calls.call_count == 0
+    assert get_user_data_calls.called
+    assert get_user_data_calls.call_count == 5
 
 
 @requests_mock.mock(kw='mocker')
 def test_inject_logged_user_when_token_is_invalid(mocker):
-    get_user_data_calls = mocker.get('http://gateway-url.com/auth/user/me',
-                                     request_headers={'Authorization': 'Bearer abcd'},
-                                     status_code=401,
-                                     json={
-                                         "errors": [
-                                             {
-                                                 "status": 401,
-                                                 "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."
-                                             }
-                                         ]
-                                     })
+    get_user_data_calls = mock_request_validation_invalid_token(mocker)
 
     test_endpoints = Blueprint('rw_api', __name__)
 
@@ -158,28 +123,34 @@ def test_inject_logged_user_when_token_is_invalid(mocker):
     RWAPIMicroservicePython.register(
         app=app,
         gateway_url='http://gateway-url.com',
-        token='microserviceToken'
+        token='microserviceToken',
+        require_api_key=True
     )
 
-    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
-    assert response.json == json.loads(b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
+    assert response.json == json.loads(
+        b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
-    assert response.json == json.loads(b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
+    assert response.json == json.loads(
+        b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
-    assert response.json == json.loads(b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
+    assert response.json == json.loads(
+        b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
-    assert response.json == json.loads(b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
+    assert response.json == json.loads(
+        b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
-    assert response.json == json.loads(b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
+    assert response.json == json.loads(
+        b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
     assert get_user_data_calls.called
     assert get_user_data_calls.call_count == 5